src/requirements.txt
transformers
torch
librosa
speechbrain
openai-whisper
